requirements.txt
requests==2.31.0
streamlit==1.34.0
python-dotenv==1.0.1
faiss-cpu==1.8.0
pypdf==4.2.0
numpy==1.26.4
scikit-learn==1.4.2
